api-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4
pydantic-settings>=2.10.1
python-dotenv==1.0.0
psutil==5.9.6
httpx==0.25.2
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.26.2
huggingface-hub>=0.20.0
pymongo==4.6.0
requests>=2.32.5
azure-ai-inference==1.0.0b9
azure-identity==1.15.0
openai==1.3.0
langchain>=0.3.27
langchain-core>=0.3.76
langchain-community>=0.3.29
structlog>=23.2.0
